requirements.txt
fastapi==0.112.2
httpx==0.27.2
pydantic==2.8.2
pytest==8.3.2
python-dotenv==1.0.1
setuptools==58.1.0
pipreqs==0.5.0
httpx[http2]==0.27.2
pytest-asyncio==0.24.0